List or Search City Names

List all city names containing a specific string.

Usage

list_cities(match = NULL)

Arguments

match

string to match.

Value

A data frame containing all matched cities alongside their respective country and coordinates. A row of the output can be used as input for the name argument in the cityview() function.

See also

Author

Koen Derks, koen-derks@hotmail.com

Examples

# List all cities containing "Ams"
list_cities(match = "Ams")
#>              name         country   lat   long
#> 1      Amstelveen The Netherlands 52.32   4.86
#> 2       Amsterdam The Netherlands 52.37   4.89
#> 3       Amstetten         Austria 48.13  14.86
#> 4   New Amsterdam          Guyana  6.25 -57.53
#> 5 Nieuw Amsterdam        Suriname  5.91 -55.07