Specify a New City

Define a city using country, latitude and longitude coordinates. The returned object can be used a input for name argument the cityview() function.

Usage

new_city(name = NULL, country = NULL, lat = NULL, long = NULL)

Arguments

name

A single string to be used as the city name.

country

A single string to be used as the country.

lat

A single numeric value to be used as the latitude.

long

A single numeric value to be used as the longitude.

Value

a data frame containing the new city alongside its respective country and coordinates.

See also

Author

Koen Derks, koen-derks@hotmail.com

Examples

city <- new_city(
  name = "Lagos", country = "Portugal",
  lat = 37.10, long = -8.68
)
#> Discovered the city of Lagos, Portugal at 37.1° / -8.68°!
if (FALSE) { # \dontrun{
cityview(name = city)
} # }