Create an Aerial City View

Create an aerial city view.

cityview(name = NULL,
         zoom = 1,
         theme = c(
           "vintage", "colored", "delftware", "lichtenstein",
           "original", "rouge", "modern"
         ),
         border = c(
           "none", "circle", "rhombus", "square",
           "hexagon", "octagon", "decagon"
         ),
         halftone = c("none", "light", "modern"),
         places = 0,
         legend = FALSE,
         filename = NULL,
         verbose = TRUE,
         license = TRUE,
         bot = FALSE)

Arguments

name

a character specifying the name of the city as provided by list_cities(). If NULL (default), chooses a random city.

zoom

a numeric value specifying the amount of zoom. Values > 1 increase zoom and values < 1 decrease zoom. The zoom can be used to speed up rendering of large cities.

theme

a character specifying the theme of the plot, or a named list specifying a custom theme (see the details section for more information about the composition of this list). Possible pre-specified themes are vintage (default), colored, delftware, lichtenstein, original, rouge and modern.

border

a character specifying the type of border to use. Possible options are none (default), circle, rhombus, square, hexagon (6 vertices), octagon (8 vertices) and decagon (10 vertices).

halftone

a character specifying the type of halftone to use. Possible options are none, light (white dither) and dark (black dither).

places

an integer specifying how many suburb, quarter and neighbourhood names to add to the image.

legend

logical. Whether to add a distance measurer and a compass in the bottom left corner of the image.

filename

character. If specified, the function exports the plot at an appropriate size and does not return a ggplot2 object.

verbose

logical. Whether to show a progress bar during execution.

license

logical. Whether to add the OpenStreetMap licence to the plot.

bot

logical. Enable functionality used by the Twitter bot.

Details

The theme argument can take a custom list as input. This list must contain the following elements:

colors

background: A color for the background.

water: A color for the water.

landuse: A color or vector of colors for the landuse.

contours: A color for the contours of landuse and buildings.

streets: A color for the streets.

buildings: A color or vector of colors for the buildings.

text: A color for the text.

font

family: the family of the font

face: the face of the font.

append: Optional. A string to append the city name at both sides.

size

borders: the family of the font

streets: the face of the font.

See also

Author

Koen Derks, koen-derks@hotmail.com

Examples

if (FALSE) {
# Create a vintage city view of Amsterdam in a circle
cityview(name = "Amsterdam", theme = "vintage", border = "circle")

# Use a custom theme, for example black and white
myTheme <- list(
  colors = list(
    background = "white",
    water = "black",
    landuse = "white",
    contours = "black",
    streets = "black",
    buildings = "white",
    text = "black",
    waterlines = "black"
  ),
  font = list(
    family = "serif",
    face = "bold"
  ),
  size = list(
    borders = list(
      contours = 0.3,
      water = 0.4,
      canal = 0.5,
      river = 0.6
    ),
    streets = list(
      path = 0.1,
      residential = 0.4,
      structure = 0.5,
      tertiary = 0.75,
      secondary = 0.8,
      highway = 0.9,
      motorway = 1,
      rails = 0.75,
      runway = 3
    )
  )
)
cityview(name = "Amsterdam", theme = myTheme)
}