Create an Aerial City View

Create an aerial city view.

cityview(name, zoom = 1,
         theme = c("original", "light", "dark", "colored", "rouge",
                   "verde", "neon", "atlantis", "antique", "lichtenstein"),
         border = c("none", "circle", "rhombus", "square",
                    "hexagon", "octagon", "decagon"),
         places = FALSE, halftone = FALSE,
         filename = NULL, verbose = TRUE,
         license = TRUE, bot = FALSE)

Arguments

name

a character specifying the name of the city as provided by list_cities().

zoom

a numeric value specifying the amount of zoom. Values > 1 increase zoom and values < 1 decrease zoom. The zoom can be used to speed up rendering of large cities.

theme

a character specifying the theme of the plot. Possible options are original, light, dark, colored, rouge, verde, neon, atlantis, antique and lichtenstein.

border

a character specifying the type of border to use. Possible options are none, circle, rhombus, square, hexagon (6 vertices), octagon (8 vertices), and decagon (10 vertices).

places

logical. Whether to add neighborhood names to the plot.

halftone

logical. Whether to dither the plot using a halftone pattern.

filename

character. If specified, the function exports the plot at an appropriate size and does NOT return a ggplot2 object.

verbose

logical. Whether to show a progress bar during execution.

license

logical. Whether to add the OpenStreetMap licence to the plot.

bot

logical. Choose automatically between cities with the same name and add a copyright licence to the image. Primarily used by the twitter bot.

See also

Author

Koen Derks, koen-derks@hotmail.com

Examples

if (FALSE) {
# Create a city view of Amsterdam in a circle
cityview(name = "Amsterdam", theme = "original", border = "circle")
}