R City Views

rcityviews is an homage to the programming language R, open-source geographic data and the art of map making. It provides code and examples to render customizable stylized city maps using data from OpenStreetMap.

For documentation on rcityviews itself, including the manual and user guide for the package, worked examples, and other tutorial information visit the package website.

Author

Koen Derks (maintainer, author)	<koen-derks@hotmail.com>

Please use the citation provided by R when citing this package. A BibTex entry is available from citation("rcityviews").

See also